配比通知单列表添加，二维码功能生成添加，图片显示修改，任务单浇筑令功能第二次修改修改

+++ app/src/main/java/com/shtoone/shtw/bean/TaskListImpQueryFragmenData.java
@@ -99,7 +99,7 @@ public class TaskListImpQueryFragmenData {
         private String kaipanriqi;
         private String org_code;
         private String guid;
-        private int shijifangliang;
+        private double shijifangliang;
         private Object sgphbguid;
         private String shejifangliang;
         private String zhuangtai;
@@ -329,11 +329,11 @@ public class TaskListImpQueryFragmenData {
             this.guid = guid;
         }
 
-        public int getShijifangliang() {
+        public double getShijifangliang() {
             return shijifangliang;
         }
 
-        public void setShijifangliang(int shijifangliang) {
+        public void setShijifangliang(double shijifangliang) {
             this.shijifangliang = shijifangliang;
         }
